code refactoring
reorganized error handlers.
