+++ src/error/mod.rs
@@ -1,1 +1,33 @@
-pub mod db;
+use crate::failure::Fail;
+use actix_web::http::StatusCode;
+use actix_web::{HttpResponse, ResponseError};
+
+// Custom error handler for failed db transactions.
+#[derive(Fail, Debug)]
+pub enum Db {
+    #[fail(display = "failed transaction")]
+    FailedTransaction,
+}
+
+impl ResponseError for Db {
+    fn error_response(&self) -> HttpResponse {
+        match *self {
+            Db::FailedTransaction => HttpResponse::new(StatusCode::INTERNAL_SERVER_ERROR),
+        }
+    }
+}
+
+/// Custom error handler for API requests.
+#[derive(Fail, Debug)]
+pub enum Oauth {
+    #[fail(display = "bad request")]
+    BadRequest,
+}
+
+impl ResponseError for Oauth {
+    fn error_response(&self) -> HttpResponse {
+        match *self {
+            Oauth::BadRequest => HttpResponse::new(StatusCode::BAD_REQUEST),
+        }
+    }
+}
